// packages/machines/date-time-picker/src/date-time-picker.types.ts
export type DateView = "day" | "month" | "year"

export interface ValueChangeDetails {
  value: DateValue[]
  valueAsString: string[]
  view: DateView
}

export interface FocusChangeDetails extends ValueChangeDetails {
  focusedValue: DateValue
  view: DateView
}

export interface ViewChangeDetails {
  view: DateView
}

export interface OpenChangeDetails {
  open: boolean
}

/* -----------------------------------------------------------------------------
 * Machine context
 * -----------------------------------------------------------------------------*/

export type SelectionMode = "single" | "multiple" | "range"

export interface IntlTranslations {
  placeholder: (locale: string) => { year: string; month: string; day: string; hour: string; min: string }
}

export type ElementIds = Partial<{
  root: string
  table(id: string): string
  tableHeader(id: string): string
  tableBody(id: string): string
  tableRow(id: string): string
  content: string
  cellTrigger(id: string): string
  prevTrigger(view: DateView): string
  nextTrigger(view: DateView): string
  viewTrigger(view: DateView): string
  clearTrigger: string
  control: string
  input(index: number): string
  trigger: string
  monthSelect: string
  yearSelect: string
  positioner: string
}>

interface PublicContext extends DirectionProperty, CommonProperties {
  /**
   * The locale (BCP 47 language tag) to use when formatting the date.
   */
  locale: string
  /**
   * The localized messages to use.
   */
  translations?: IntlTranslations
  /**
   * The ids of the elements in the date picker. Useful for composition.
   */
  ids?: ElementIds
  /**
   * The `name` attribute of the input element.
   */
  name?: string
  /**
   * The time zone to use
   */
  timeZone: string
  /**
   * Whether the calendar is disabled.
   */
  disabled?: boolean
  /**
   * Whether the calendar is read-only.
   */
  readOnly?: boolean
  /**
   * The minimum date that can be selected.
   */
  min?: DateValue
  /**
   * The maximum date that can be selected.
   */
  max?: DateValue
  /**
   * Whether the calendar should close after the date selection is complete.
   * This is ignored when the selection mode is `multiple`.
   * @default true
   */
  closeOnSelect?: boolean
  /**
   * The selected date(s).
   */
  value: DateValue[]
  /**
   * The focused date.
   */
  focusedValue: DateValue
  /**
   * The number of months to display.
   */
  numOfMonths: number
  /**
   * The first day of the week.
   *  `0` - Sunday
   *  `1` - Monday
   *  `2` - Tuesday
   *  `3` - Wednesday
   *  `4` - Thursday
   *  `5` - Friday
   *  `6` - Saturday
   */
  startOfWeek?: number
  /**
   * Whether the calendar should have a fixed number of weeks.
   * This renders the calendar with 6 weeks instead of 5 or 6.
   */
  fixedWeeks?: boolean
  /**
   * Function called when the value changes.
   */
  onValueChange?: (details: ValueChangeDetails) => void
  /**
   * Function called when the focused date changes.
   */
  onFocusChange?: (details: FocusChangeDetails) => void
  /**
   * Function called when the view changes.
   */
  onViewChange?: (details: ViewChangeDetails) => void
  /**
   * Function called when the calendar opens or closes.
   */
  onOpenChange?: (details: OpenChangeDetails) => void
  /**
   * Returns whether a date of the calendar is available.
   */
  isDateUnavailable?: (date: DateValue, locale: string) => boolean
  /**
   * The selection mode of the calendar.
   * - `single` - only one date can be selected
   * - `multiple` - multiple dates can be selected
   * - `range` - a range of dates can be selected
   */
  selectionMode: SelectionMode
  /**
   * The format of the date to display in the input.
   */
  format?: (date: DateValue) => string
  /**
   * The format of the date to display in the input.
   */
  parse?: (value: string) => DateValue
  /**
   * The view of the calendar
   * @default "day"
   */
  view: DateView
  /**
   * Whether the calendar should be modal. This means that the calendar will
   * block interaction with the rest of the page, and trap focus within it.
   */
  modal?: boolean
  /**
   * The user provided options used to position the date picker content
   */
  positioning: PositioningOptions
  /**
   * Whether the datetimepicker is open
   */
  open?: boolean
  /**
   * Whether the datetimepicker open state is controlled by the user
   */
  "open.controlled"?: boolean
}

interface PrivateContext {
  /**
   * @internal
   * The start date of the current visible duration.
   */
  startValue: DateValue
  /**
   * @internal
   * Whether the calendar has focus
   */
  hasFocus?: boolean
  /**
   * @internal
   * The live region to announce changes
   */
  announcer?: LiveRegion
  /**
   * @internal
   * The current hovered date. Useful for range selection mode.
   */
  hoveredValue: DateValue | null
  /**
   * @internal
   * The index of the currently active date.
   * Used in range selection mode.
   */
  activeIndex: number
  /**
   * @internal
   * The computed placement (maybe different from initial placement)
   */
  currentPlacement?: Placement
  /**
   * @internal
   * Whether the calendar should restore focus to the input when it closes.
   */
  restoreFocus?: boolean
}

type ComputedContext = Readonly<{
  /**
   * @computed
   * The end date of the current visible duration.
   */
  endValue: DateValue
  /**
   * @computed
   * Whether the calendar is interactive.
   */
  isInteractive: boolean
  /**
   * @computed
   * The duration of the visible range.
   */
  visibleDuration: DateDuration
  /**
   * @computed
   * The start/end date of the current visible duration.
   */
  visibleRange: { start: DateValue; end: DateValue }
  /**
   * @computed
   * The text to announce when the visible range changes.
   */
  visibleRangeText: { start: string; end: string; formatted: string }
  /**
   * @computed
   * Whether the next visible range is valid.
   */
  isNextVisibleRangeValid: boolean
  /**
   * @computed
   * Whether the previous visible range is valid.
   */
  isPrevVisibleRangeValid: boolean
  /**
   * @computed
   * The value text to display in the input.
   */
  valueAsString: string[]
  /**
   * @internal
   * The input element's value
   */
  formattedValue: string[]
}>

export type UserDefinedContext = RequiredBy<PublicContext, "id">

export interface MachineContext extends PublicContext, PrivateContext, ComputedContext {}